+++ packages/components/src/interface/index.ts
@@ -20,7 +20,6 @@ export * from './SearchTool'
 export * from './ExpandableNotification/ExpandableNotification'
 export * from './Header/AppHeader'
 export * from './Header/PageHeader'
-export * from './ToggleMenu'
 export * from './ViewData'
 export * from './RadioButton'
 export * from './EventTopBar'
